requirements.txt
# Core Deep Learning
torch>=2.1.0,<2.5.0
torchvision>=0.16.0,<0.20.0
torchaudio>=2.1.0,<2.5.0

# Video & Image Processing
opencv-python>=4.8.0,<4.11.0
mediapipe>=0.10.7,<0.11.0
av>=11.0.0,<13.0.0
albumentations>=1.3.1,<1.5.0

# Numerical & Data
numpy>=1.24.0,<2.1.0
pandas>=2.0.0,<2.3.0
scikit-learn>=1.3.0,<1.6.0

# Visualization
matplotlib>=3.7.0,<3.10.0
seaborn>=0.12.0,<0.14.0

# Experiment Tracking
wandb>=0.16.0,<0.19.0
mlflow>=2.9.0,<2.18.0
tensorboard>=2.14.0,<2.19.0

# Model Export & Optimization
onnx>=1.15.0,<1.18.0
onnxruntime>=1.16.0,<1.20.0

# Deep Learning Utilities
einops>=0.7.0,<0.9.0
timm>=0.9.10,<1.1.0
pytorchvideo>=0.1.5,<0.2.0

# Configuration & CLI
PyYAML>=6.0.1,<6.1.0
rich>=13.0.0,<13.10.0

# Progress Bars
tqdm>=4.66.0,<4.68.0

# Dataset Download
gdown>=5.0.0,<5.3.0
kaggle>=1.6.0,<1.8.0

# Jupyter Notebooks
jupyter>=1.0.0,<1.2.0
ipywidgets>=8.0.0,<8.3.0
